+++ src/pages/CancelledPage/CancelledPage.tsx
@@ -1,7 +1,9 @@
+//  IMPORTS
+
 import { Header } from "../header/header";
 import { Link } from "react-router-dom";
 import { Address } from "../../components/address/address";
-import "./styles/cancelledBooking.css";
+import "./styles/cancelledPage.css";
 
 export const CancelledBooking = () => {
   return (
