+++ src/pages/CancelledPage/styles/cancelledPage.css
@@ -35,4 +35,4 @@
     width: 100%;
     font-size: 22pt;
   }
-}/*# sourceMappingURL=cancelledBooking.css.map */
+}/*# sourceMappingURL=cancelledPage.css.map */
