+++ src/pages/CancelledPage/styles/cancelledPage.scss
